chore: move `formatLatLon` from `geo.ts` to `format.ts`
